+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Feat: Support envelope based transport for events (#391)
 * Feature: Envelope Only Transport API #426 (#463)
   * The method signature of `Transport` changed from `Future<SentryId> send(SentryEvent event)` to `Future<SentryId> send(SentryEnvelope envelope)`
+* Fix: Merge user from event and scope (#467)
 
 # 5.1.0-beta.1